Icon and Name App

+++ lib/widgets/chart_item_widget.dart
@@ -98,7 +98,7 @@ class _ChartItemWidgetState extends State<ChartItemWidget> {
     );
   }
 
-  double getPrice() {
+  int getPrice() {
     return widget.item!.price! * amount;
   }
 }
